refactor(architecture): converge the shell wave and empty the waiver ledger (MOD-DASHBOARD-SHELL)
Three structural cuts close the final wave:

1. isUriString converges into the kernel (src/uriStrings.ts) — three local
   copies (openProjectService, openProjectMatcher, gitRepositoryDetector)
   collapse into one owner, and the RUNTIME->PROJECT-CATALOG edge dies with
   it, dissolving the last SCC cluster. The cycle baseline and the waiver
   ledger are both empty.
2. The provider executable probe (resolveAiProviderExecutable,
   runBoundedAiProviderHelp) moves from dashboard.ts to the provider
   directory capability that already defines their adapter interface.
3. The vscode.git API acquisition moves to src/dashboard/gitApiAcquisition.ts
   — deliberately shell-internal, because the worktree module must stay
   loadable without the vscode runtime (its unit tests load it bare).

dashboard.ts sheds its three top-level business helpers (3993 -> 3927 lines).
Shell entrypoints narrow to the composition root plus the five consumed type
surfaces; sponsor.ts becomes module-internal. Ledger flips
MOD-DASHBOARD-SHELL to strict: every module is now strict.
